chore: refactor with clean up watch/mod.rs
