+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod watch;
 use config::CommandType;
 use log::error;
 use std::error::Error;
+use watch::watch::watch;
 
 pub async fn run(
     dir: String,
@@ -15,7 +16,7 @@ pub async fn run(
     bin_arg: Option<Vec<String>>,
 ) -> Result<(), Box<dyn Error + Send>> {
     if cfg!(not(test)) {
-        if let Err(err) = watch::watch(dir, cmd, ignore, bin_path, bin_arg).await {
+        if let Err(err) = watch(dir, cmd, ignore, bin_path, bin_arg).await {
             error!("Error watching directory: {:?}", err)
         }
     }
