+++ src/watch/command.rs
@@ -1,9 +1,13 @@
-use std::process::{Child, Stdio};
+use std::{
+    io::{BufRead, BufReader},
+    process::{Child, Command, Stdio},
+};
+use tokio::task;
 
 pub async fn exec(cmd: String) -> Result<Child, Box<dyn std::error::Error>> {
-    let output = tokio::task::spawn_blocking(move || {
+    let output = task::spawn_blocking(move || {
         if cfg!(windows) {
-            std::process::Command::new("cmd")
+            Command::new("cmd")
                 .arg("/C")
                 .arg(cmd)
                 .stdout(Stdio::piped())
@@ -11,7 +15,7 @@ pub async fn exec(cmd: String) -> Result<Child, Box<dyn std::error::Error>> {
                 .spawn()
                 .unwrap()
         } else {
-            std::process::Command::new("sh")
+            Command::new("sh")
                 .arg("-c")
                 .arg(cmd)
                 .stdout(Stdio::piped())
@@ -23,6 +27,24 @@ pub async fn exec(cmd: String) -> Result<Child, Box<dyn std::error::Error>> {
     .await?;
 
     Ok(output)
+}
+
+pub fn cmd_result(child: Child) {
+    macro_rules! process_output {
+        ($reader:expr, $print_fn:ident) => {
+            for line in $reader.lines() {
+                $print_fn!("{}", line.unwrap());
+            }
+        };
+    }
+
+    let stdout = child.stdout.unwrap();
+    let stderr = child.stderr.unwrap();
+    let stdout_reader = BufReader::new(stdout);
+    let stderr_reader = BufReader::new(stderr);
+
+    process_output!(stdout_reader, println);
+    process_output!(stderr_reader, eprintln);
 }
 
 #[cfg(test)]
